support_18988_change_dockerfile_to_multistage_build
First iteration of the solution using docker build caching technique.

Running for the first time will require a download of the dependecies, however the next runs will not. Only when the POM is changed will dependecies be downloaded again.

the docker build process does not support the use of mounted volumes which is why this technique is being used currently